pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "crittercam"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "pillow",
    "speciesnet",
    "timm",
    "tomli-w",
    "uvicorn",
]

[project.scripts]
crittercam = "crittercam.cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["crittercam*"]
